requirements.txt
flask
flask-socketio
eventlet
pymongo
Flask-PyMongo
dnspython
python-dotenv
google-generativeai
PyJWT
itsdangerous
anthropic
flask-cors
langchain
openai
langchain-community
langchain-openai
chromadb
sentence-transformers
tiktoken
openrouter
openai-whisper
torch
ffmpeg-python
boto3
Pillow
pillow-heif
mem0ai
chromadb
qdrant-client
google-genai